+++ lib/Imagine/Filter/Advanced/Grayscale.php
@@ -12,7 +12,6 @@
 namespace Imagine\Filter\Advanced;
 
 use Imagine\Filter\FilterInterface;
-use Imagine\Image\Color;
 use Imagine\Image\ImageInterface;
 use Imagine\Image\Point;
 
@@ -25,13 +24,7 @@ class Grayscale extends OnPixelBased implements FilterInterface
     {
         parent::__construct(function (ImageInterface $image, Point $point) {
             $color = $image->getColorAt($point);
-            $gray  = min(255, round(($color->getRed() + $color->getBlue() + $color->getGreen())/3));
-
-            $image->draw()->dot($point, new Color(array(
-                'red'   => $gray,
-                'green' => $gray,
-                'blue'  => $gray
-            )));
+            $image->draw()->dot($point, $color->grayscale());
         });
     }
 }