+++ lib/Imagine/Imagick/Drawer.php
@@ -16,7 +16,7 @@ use Imagine\Exception\InvalidArgumentException;
 use Imagine\Exception\RuntimeException;
 use Imagine\Image\AbstractFont;
 use Imagine\Image\BoxInterface;
-use Imagine\Image\Color;
+use Imagine\Image\Palette\Color\ColorInterface;
 use Imagine\Image\Point;
 use Imagine\Image\PointInterface;
 
@@ -41,7 +41,7 @@ final class Drawer implements DrawerInterface
     /**
      * {@inheritdoc}
      */
-    public function arc(PointInterface $center, BoxInterface $size, $start, $end, Color $color, $thickness = 1)
+    public function arc(PointInterface $center, BoxInterface $size, $start, $end, ColorInterface $color, $thickness = 1)
     {
         $x      = $center->getX();
         $y      = $center->getY();
@@ -76,7 +76,7 @@ final class Drawer implements DrawerInterface
     /**
      * {@inheritdoc}
      */
-    public function chord(PointInterface $center, BoxInterface $size, $start, $end, Color $color, $fill = false, $thickness = 1)
+    public function chord(PointInterface $center, BoxInterface $size, $start, $end, ColorInterface $color, $fill = false, $thickness = 1)
     {
         $x      = $center->getX();
         $y      = $center->getY();
@@ -136,7 +136,7 @@ final class Drawer implements DrawerInterface
     /**
      * {@inheritdoc}
      */
-    public function ellipse(PointInterface $center, BoxInterface $size, Color $color, $fill = false, $thickness = 1)
+    public function ellipse(PointInterface $center, BoxInterface $size, ColorInterface $color, $fill = false, $thickness = 1)
     {
         $width  = $size->getWidth();
         $height = $size->getHeight();
@@ -183,7 +183,7 @@ final class Drawer implements DrawerInterface
     /**
      * {@inheritdoc}
      */
-    public function line(PointInterface $start, PointInterface $end, Color $color, $thickness = 1)
+    public function line(PointInterface $start, PointInterface $end, ColorInterface $color, $thickness = 1)
     {
         try {
             $pixel = $this->getColor($color);
@@ -218,7 +218,7 @@ final class Drawer implements DrawerInterface
     /**
      * {@inheritdoc}
      */
-    public function pieSlice(PointInterface $center, BoxInterface $size, $start, $end, Color $color, $fill = false, $thickness = 1)
+    public function pieSlice(PointInterface $center, BoxInterface $size, $start, $end, ColorInterface $color, $fill = false, $thickness = 1)
     {
         $width  = $size->getWidth();
         $height = $size->getHeight();
@@ -252,7 +252,7 @@ final class Drawer implements DrawerInterface
     /**
      * {@inheritdoc}
      */
-    public function dot(PointInterface $position, Color $color)
+    public function dot(PointInterface $position, ColorInterface $color)
     {
         $x = $position->getX();
         $y = $position->getY();
@@ -283,7 +283,7 @@ final class Drawer implements DrawerInterface
     /**
      * {@inheritdoc}
      */
-    public function polygon(array $coordinates, Color $color, $fill = false, $thickness = 1)
+    public function polygon(array $coordinates, ColorInterface $color, $fill = false, $thickness = 1)
     {
         if (count($coordinates) < 3) {
             throw new InvalidArgumentException(sprintf(
@@ -389,13 +389,13 @@ final class Drawer implements DrawerInterface
     }
 
     /**
-     * Gets specifically formatted color string from Color instance
+     * Gets specifically formatted color string from ColorInterface instance
      *
-     * @param Color $color
+     * @param ColorInterface $color
      *
      * @return string
      */
-    private function getColor(Color $color)
+    private function getColor(ColorInterface $color)
     {
         $pixel = new \ImagickPixel((string) $color);